# specs/002-pdp-product/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: PDP P1 平台 API
  version: 1.1.0
  description: |
    PDP 企业级项目交付管理平台 P1 的稳定同步接口边界。
    本契约与 coverage.md 共同覆盖全部 P1 稳定操作；领域包生成页面使用
    domain-package.schema.json 定义的元数据，不为客户复制私有接口。
...
servers:
  - url: /api/v1
security:
  - oidc: []
...
paths:
  /workspaces:
    get:
      tags: [工作空间]
      summary: 获取当前用户可访问的工作空间
      operationId: listWorkspaces
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          $ref: '#/components/responses/WorkspacePage'
    post:
      tags: [工作空间]
      summary: 创建工作空间
      operationId: createWorkspace
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkspaceCommand'
      responses:
        '201':
          description: 已创建
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
        '409':
          $ref: '#/components/responses/Problem'
  /workspaces/{workspaceId}:
    get:
      tags: [工作空间]
      summary: 获取工作空间治理详情
      operationId: getWorkspace
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdPath'
      responses:
        '200':
          description: 工作空间详情
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
        '404':
          $ref: '#/components/responses/Problem'
  /workspaces/{workspaceId}/actions:
    post:
      tags: [工作空间]
      summary: 执行工作空间状态迁移
      operationId: applyWorkspaceAction
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdPath'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkspaceActionCommand'
      responses:
        '200':
          description: 状态迁移成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
        '409':
          $ref: '#/components/responses/Problem'
  /workspaces/{workspaceId}/organizations:
    get:
      tags: [工作空间]
      summary: 查询工作空间组织结构
      operationId: listOrganizationUnits
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdPath'
      responses:
        '200':
          description: 组织单元
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrganizationUnit'
    post:
      tags: [工作空间]
      summary: 创建组织单元
      operationId: createOrganizationUnit
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdPath'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrganizationUnitCommand'
      responses:
        '201':
          description: 组织单元已创建
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationUnit'
        '409':
          $ref: '#/components/responses/Problem'
  /workspaces/{workspaceId}/members:
    get:
      tags: [工作空间]
      summary: 查询工作空间成员
      operationId: listWorkspaceMembers
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdPath'
      responses:
        '200':
          description: 成员列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkspaceMembership'
    post:
      tags: [工作空间]
      summary: 添加工作空间成员
      operationId: addWorkspaceMember
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdPath'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddWorkspaceMemberCommand'
      responses:
        '201':
          description: 成员已添加
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceMembership'
        '409':
          $ref: '#/components/responses/Problem'
  /workspaces/{workspaceId}/roles:
    get:
      tags: [工作空间]
      summary: 查询工作空间角色
      operationId: listWorkspaceRoles
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdPath'
      responses:
        '200':
          description: 角色列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkspaceRole'
    post:
      tags: [工作空间]
      summary: 创建工作空间角色
      operationId: createWorkspaceRole
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdPath'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkspaceRoleCommand'
      responses:
        '201':
          description: 角色已创建
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceRole'
        '409':
          $ref: '#/components/responses/Problem'
  /workspaces/{workspaceId}/data-scopes:
    post:
      tags: [工作空间]
      summary: 创建数据范围
      operationId: createDataScope
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdPath'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDataScopeCommand'
      responses:
        '201':
          description: 数据范围已创建
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataScope'
        '409':
          $ref: '#/components/responses/Problem'
  /workspaces/{workspaceId}/collaboration-grants:
    get:
      tags: [工作空间]
      summary: 查询跨工作空间协作授权
      operationId: listCollaborationGrants
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdPath'
      responses:
        '200':
          description: 授权列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CollaborationGrant'
    post:
      tags: [工作空间]
      summary: 授予跨工作空间协作权限
      operationId: createCollaborationGrant
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdPath'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCollaborationGrantCommand'
      responses:
        '201':
          description: 授权已创建
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollaborationGrant'
        '403':
          $ref: '#/components/responses/Problem'
  /workspaces/{workspaceId}/collaboration-grants/{grantId}/revoke:
    post:
      tags: [工作空间]
      summary: 撤销跨工作空间授权
      operationId: revokeCollaborationGrant
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdPath'
        - name: grantId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Uuid'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReasonCommand'
      responses:
        '204':
          description: 已撤销
        '409':
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://idp.example.com/.well-known/openid-configuration
  parameters:
...
    WorkspaceIdPath:
      name: workspaceId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Uuid'
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 128
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: 当前资源 revision 对应的强校验值；不匹配返回 409
      schema:
        type: string
    Cursor:
      name: cursor
      in: query
      description: 平台签发的不透明 keyset 游标，绑定排序、过滤和权限范围；禁止客户端解析或修改
      schema:
        type: string
        minLength: 16
        maxLength: 2048
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
  responses:
    Problem:
      description: 标准问题响应
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    WorkspacePage:
      description: 工作空间分页
      content:
        application/json:
          schema:
            type: object
            required: [items, nextCursor]
            properties:
              items:
                type: array
                items:
                  $ref: '#/components/schemas/Workspace'
              nextCursor:
                type: [string, 'null']
                description: 下一页不透明 keyset 游标；为 null 表示没有下一页
  schemas:
    Uuid:
      type: string
      format: uuid
    ObjectRef:
      type: object
      required: [objectType, objectId]
      properties:
        objectType:
          type: string
          pattern: '^[a-z][a-z0-9.-]{1,99}$'
        objectId:
          $ref: '#/components/schemas/Uuid'
...
    Problem:
      type: object
      required: [type, title, status, traceId]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        code:
          type: string
        traceId:
          type: string
        violations:
          type: array
          items:
            type: object
            required: [field, code]
            properties:
              field:
                type: string
              code:
                type: string
              message:
                type: string
    CreateWorkspaceCommand:
      type: object
      required: [code, name, ownerUserId]
      properties:
        code:
          type: string
          pattern: '^[A-Z][A-Z0-9_-]{2,31}$'
        name:
          type: string
          minLength: 2
          maxLength: 100
        ownerUserId:
          $ref: '#/components/schemas/Uuid'
        defaultLocale:
          type: string
          default: zh-CN
        defaultTimezone:
          type: string
          default: Asia/Shanghai
    Workspace:
      type: object
      required: [id, code, name, ownerUserId, status, defaultLocale, defaultTimezone, revision]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        code:
          type: string
        name:
          type: string
        ownerUserId:
          $ref: '#/components/schemas/Uuid'
        status:
          type: string
          enum: [DRAFT, ACTIVE, SUSPENDED, ARCHIVED]
        revision:
          type: integer
          minimum: 0
    WorkspaceActionCommand:
      type: object
      required: [action, reason]
      properties:
        action:
          type: string
          enum: [ACTIVATE, SUSPEND, ARCHIVE]
        reason:
          type: string
          minLength: 5
          maxLength: 500
    CreateOrganizationUnitCommand:
      type: object
      required: [code, name, type]
      properties:
        parentId:
          oneOf:
            - $ref: '#/components/schemas/Uuid'
            - type: 'null'
        code:
          type: string
          pattern: '^[A-Z][A-Z0-9_-]{1,31}$'
        name:
          type: string
          minLength: 2
          maxLength: 100
        type:
          type: string
          enum: [COMPANY, DEPARTMENT, TEAM, REGION, EXTERNAL]
        regionCode:
          type: [string, 'null']
    OrganizationUnit:
      allOf:
        - $ref: '#/components/schemas/CreateOrganizationUnitCommand'
        - type: object
          required: [id, workspaceId, path, status, revision]
          properties:
            id:
              $ref: '#/components/schemas/Uuid'
            workspaceId:
              $ref: '#/components/schemas/Uuid'
            path:
              type: string
            status:
              type: string
              enum: [ACTIVE, DISABLED]
            revision:
              type: integer
              minimum: 0
    AddWorkspaceMemberCommand:
      type: object
      required: [userId, roleIds, dataScopeIds, membershipType]
      properties:
        userId:
          $ref: '#/components/schemas/Uuid'
        organizationUnitId:
          oneOf:
            - $ref: '#/components/schemas/Uuid'
            - type: 'null'
        roleIds:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: '#/components/schemas/Uuid'
        dataScopeIds:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/Uuid'
        membershipType:
          type: string
          enum: [INTERNAL, EXTERNAL]
        validUntil:
          type: [string, 'null']
          format: date-time
    WorkspaceMembership:
      allOf:
        - $ref: '#/components/schemas/AddWorkspaceMemberCommand'
        - type: object
          required: [id, workspaceId, status, revision]
          properties:
            id:
              $ref: '#/components/schemas/Uuid'
            workspaceId:
              $ref: '#/components/schemas/Uuid'
            status:
              type: string
              enum: [INVITED, ACTIVE, SUSPENDED, EXPIRED, REVOKED]
            revision:
              type: integer
              minimum: 0
    CreateWorkspaceRoleCommand:
      type: object
      required: [stableKey, name, allowedActions]
      properties:
        stableKey:
          type: string
          pattern: '^[a-z][a-z0-9.-]{2,99}$'
        name:
          type: string
          minLength: 2
          maxLength: 100
        allowedActions:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            type: string
    WorkspaceRole:
      allOf:
        - $ref: '#/components/schemas/CreateWorkspaceRoleCommand'
        - type: object
          required: [id, workspaceId, status, revision]
          properties:
            id:
              $ref: '#/components/schemas/Uuid'
            workspaceId:
              $ref: '#/components/schemas/Uuid'
            status:
              type: string
              enum: [ACTIVE, RETIRED]
            revision:
              type: integer
              minimum: 0
    CreateDataScopeCommand:
      type: object
      required: [stableKey, name, resourceTypes, condition]
      properties:
        stableKey:
          type: string
          pattern: '^[a-z][a-z0-9.-]{2,99}$'
        name:
          type: string
          minLength: 2
          maxLength: 100
        resourceTypes:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            type: string
        condition:
          type: object
          additionalProperties: true
    DataScope:
      allOf:
        - $ref: '#/components/schemas/CreateDataScopeCommand'
        - type: object
          required: [id, workspaceId, status, revision]
          properties:
            id:
              $ref: '#/components/schemas/Uuid'
            workspaceId:
              $ref: '#/components/schemas/Uuid'
            status:
              type: string
              enum: [ACTIVE, RETIRED]
            revision:
              type: integer
              minimum: 0
    CreateCollaborationGrantCommand:
      type: object
      required: [collaboratorWorkspaceId, target, roleId, allowedActions, validUntil]
      properties:
        collaboratorWorkspaceId:
          $ref: '#/components/schemas/Uuid'
        target:
          $ref: '#/components/schemas/ObjectRef'
        roleId:
          $ref: '#/components/schemas/Uuid'
        allowedActions:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            type: string
        validUntil:
          type: string
          format: date-time
        reason:
          type: string
          maxLength: 500
    CollaborationGrant:
      allOf:
        - $ref: '#/components/schemas/CreateCollaborationGrantCommand'
        - type: object
          required: [id, status, revision]
          properties:
            id:
              $ref: '#/components/schemas/Uuid'
            status:
              type: string
              enum: [DRAFT, ACTIVE, EXPIRED, REVOKED]
            revision:
              type: integer
    ReasonCommand:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
          minLength: 2
          maxLength: 1000
...